+++ api/src/main/java/com/dumbdogdiner/stickychatapi/StickyChat.java
@@ -1,5 +1,6 @@
 package com.dumbdogdiner.stickychatapi;
 
+import com.dumbdogdiner.stickychatapi.chat.*;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.ServicePriority;
