+++ api/src/main/java/com/dumbdogdiner/stickychatapi/misc/DeathMessageService.java
@@ -1,0 +1,8 @@
+package com.dumbdogdiner.stickychatapi.misc;
+
+/**
+ * Handles the registration and distribution of death messages.
+ */
+public interface DeathMessageService {
+    get
+}
