+++ api/src/main/java/com/dumbdogdiner/stickychatapi/misc/MotdService.java
@@ -1,0 +1,4 @@
+package com.dumbdogdiner.stickychatapi.misc;
+
+public interface MotdService {
+}
